==> NetflixTitles/0_ETL/2.0_Extract.ipynb <==
import pandas as pd

df = pd.read_csv('../1_Data/raw_dataset.csv')

df.head()

df.isnull().sum()

df.dropna(subset=['director', 'cast', 'country', 'date_added', 'duration', 'rating'], inplace = True)

df.reset_index(drop=True, inplace=True)

df.head()

# use this code if you want to replace nan values with None for MySQL. It will keep the tuple but some of its data will just have Null in some columns.
# import numpy as np
# print(df.isnull().sum())
# df_titles = df_titles.replace({np.nan: None})

df.dtypes

#Save df as csv for next file
df.to_csv('../1_Data/extracted_dataset.csv', index=False)
