==> NetflixTitles/0_ETL/3.0_Transform.ipynb <==
import pandas as pd

df = pd.read_csv('../1_Data/extracted_dataset.csv')

df.head()

#Drop show_id and description columns
#Format date_added to be proper SQL date format
#Figure out how to split country, and listed_in columns since there's multiples

cols = ['show_id', 'description']

df.drop(cols, inplace = True, axis = 1)

df.head()

df['date_added'] = pd.to_datetime(df['date_added'], format='%d-%b-%y', errors='coerce')
df['date_added'] = df['date_added'].fillna(pd.to_datetime(df['date_added'], errors='coerce').dt.strftime('%Y-%m-%d'))

df.head()

df.isnull().sum()

df.dropna(subset=['date_added'], inplace = True)

df.reset_index(drop=True, inplace=True)

df.head()

df.dtypes

df_titles = df[['type', 'title', 'release_year', 'date_added', 'rating', 'duration']]

df_titles.head()

df_titles.to_csv('../1_Data/transformed_titles_dataset.csv', index=False)

# Splitting multiple values and creating separate rows
df_directors = df[['title', 'director']].assign(director=df['director'].str.split(', ')).explode('director')

df_directors.head()

df_directors.to_csv('../1_Data/transformed_directors_dataset.csv', index=False)

df_cast = df[['title', 'cast']].assign(cast=df['cast'].str.split(', ')).explode('cast')

df_cast = df_cast.rename(columns={'cast': 'actor'})

df_cast.head()

df_cast.to_csv('../1_Data/transformed_cast_dataset.csv', index=False)

df_countries = df[['title', 'country']].assign(country=df['country'].str.split(', ')).explode('country')

df_countries.head()

df_countries.to_csv('../1_Data/transformed_countries_dataset.csv', index=False)

df_genres = df[['title', 'listed_in']].assign(listed_in=df['listed_in'].str.split(', ')).explode('listed_in')

df_genres = df_genres.rename(columns={'listed_in': 'genre'})

df_genres.head()

df_genres.to_csv('../1_Data/transformed_genres_dataset.csv', index=False)
