==> NetflixTitles/0_ETL/4.0_Load.ipynb <==
import pandas as pd
import os
from dotenv import load_dotenv
import mysql.connector

load_dotenv()

configuration = {
    'user' : os.getenv('user'),
    'password' : os.getenv('password'),
    'host' : os.getenv('host'),
    'database' : 'information_schema',
    'raise_on_warnings' : True
}

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()

    cursor.execute("SELECT SCHEMA_NAME FROM SCHEMATA WHERE SCHEMA_NAME = 'netflix_titles';")
    schema_exists = cursor.fetchone()

    if not schema_exists:
        cursor.execute("CREATE SCHEMA netflix_titles;")
        print("Schema netflix_titles created.")

        cursor.execute("USE netflix_titles;")

    else:
        print("Schema already exists.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()
    
    cursor.execute("""
                    SELECT EXISTS (
                    SELECT TABLE_NAME
                    FROM information_schema.TABLES
                    WHERE TABLE_SCHEMA LIKE 'netflix_titles' AND TABLE_TYPE LIKE 'BASE TABLE' AND TABLE_NAME = 'titles');
                """)
    table_exists = cursor.fetchone()[0]
    
    if not table_exists:
        cursor.execute("USE netflix_titles;")
        cursor.execute("""
                    CREATE TABLE titles (
                    titles_id INT AUTO_INCREMENT PRIMARY KEY,
                    type VARCHAR(255) NOT NULL,
                    title VARCHAR(255) NOT NULL,
                    release_year YEAR,
                    date_added DATE,
                    rating VARCHAR(50),
                    duration VARCHAR(50)
                    );
                """)
        print("The 'titles' table has been created.")
        
    else:
        print("The 'titles' table already exists.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()
    
    cursor.execute("""
                    SELECT EXISTS (
                    SELECT TABLE_NAME
                    FROM information_schema.TABLES
                    WHERE TABLE_SCHEMA LIKE 'netflix_titles' AND TABLE_TYPE LIKE 'BASE TABLE' AND TABLE_NAME = 'directors');
                """)
    table_exists = cursor.fetchone()[0]
    
    if not table_exists:
        cursor.execute("USE netflix_titles;")
        cursor.execute("""
                    CREATE TABLE directors (
                    director_id INT AUTO_INCREMENT PRIMARY KEY,
                    title VARCHAR(255) NOT NULL,
                    director VARCHAR(255) NOT NULL
                    );
                """)
        print("The 'directors' table has been created.")
        
    else:
        print("The 'directors' table already exists.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()
    
    cursor.execute("""
                    SELECT EXISTS (
                    SELECT TABLE_NAME
                    FROM information_schema.TABLES
                    WHERE TABLE_SCHEMA LIKE 'netflix_titles' AND TABLE_TYPE LIKE 'BASE TABLE' AND TABLE_NAME = 'cast');
                """)
    table_exists = cursor.fetchone()[0]
    
    if not table_exists:
        cursor.execute("USE netflix_titles;")
        cursor.execute("""
                    CREATE TABLE cast (
                    cast_id INT AUTO_INCREMENT PRIMARY KEY,
                    title VARCHAR(255) NOT NULL,
                    actor VARCHAR(255) NOT NULL
                    );
                """)
        print("The 'cast' table has been created.")
        
    else:
        print("The 'cast' table already exists.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()
    
    cursor.execute("""
                    SELECT EXISTS (
                    SELECT TABLE_NAME
                    FROM information_schema.TABLES
                    WHERE TABLE_SCHEMA LIKE 'netflix_titles' AND TABLE_TYPE LIKE 'BASE TABLE' AND TABLE_NAME = 'countries');
                """)
    table_exists = cursor.fetchone()[0]
    
    if not table_exists:
        cursor.execute("USE netflix_titles;")
        cursor.execute("""
                    CREATE TABLE countries (
                    country_id INT AUTO_INCREMENT PRIMARY KEY,
                    title VARCHAR(255) NOT NULL,
                    country VARCHAR(255) NOT NULL
                    );
                """)
        print("The 'countries' table has been created.")
        
    else:
        print("The 'countries' table already exists.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()
    
    cursor.execute("""
                    SELECT EXISTS (
                    SELECT TABLE_NAME
                    FROM information_schema.TABLES
                    WHERE TABLE_SCHEMA LIKE 'netflix_titles' AND TABLE_TYPE LIKE 'BASE TABLE' AND TABLE_NAME = 'genres');
                """)
    table_exists = cursor.fetchone()[0]
    
    if not table_exists:
        cursor.execute("USE netflix_titles;")
        cursor.execute("""
                    CREATE TABLE genres (
                    genre_id INT AUTO_INCREMENT PRIMARY KEY,
                    title VARCHAR(255) NOT NULL,
                    genre VARCHAR(255) NOT NULL
                    );
                """)
        print("The 'genres' table has been created.")
        
    else:
        print("The 'genres' table already exists.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

df_titles = pd.read_csv('../1_Data/transformed_titles_dataset.csv')

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()

    cursor.execute("USE netflix_titles;")
    
    insert_stmt = """
        INSERT INTO titles (type, title, release_year, date_added, rating, duration)
        VALUES (%s, %s, %s, %s, %s, %s)
    """

    # Iterate through the DataFrame and insert each row
    for _, row in df_titles.iterrows():
        cursor.execute(insert_stmt, tuple(row))

    # Commit the transaction
    connection.commit()

    print("transformed_titles_dataset.csv inserted into MySQL successfully.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

df_directors = pd.read_csv('../1_Data/transformed_directors_dataset.csv')

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()

    cursor.execute("USE netflix_titles;")
    
    insert_stmt = """
        INSERT INTO directors (title, director)
        VALUES (%s, %s)
    """

    # Iterate through the DataFrame and insert each row
    for _, row in df_directors.iterrows():
        cursor.execute(insert_stmt, tuple(row))

    # Commit the transaction
    connection.commit()

    print("transformed_directors_dataset.csv inserted into MySQL successfully.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

df_cast = pd.read_csv('../1_Data/transformed_cast_dataset.csv')

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()

    cursor.execute("USE netflix_titles;")
    
    insert_stmt = """
        INSERT INTO cast (title, actor)
        VALUES (%s, %s)
    """

    # Iterate through the DataFrame and insert each row
    for _, row in df_cast.iterrows():
        cursor.execute(insert_stmt, tuple(row))

    # Commit the transaction
    connection.commit()

    print("transformed_cast_dataset.csv inserted into MySQL successfully.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

df_countries = pd.read_csv('../1_Data/transformed_countries_dataset.csv')

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()

    cursor.execute("USE netflix_titles;")
    
    insert_stmt = """
        INSERT INTO countries (title, country)
        VALUES (%s, %s)
    """

    # Iterate through the DataFrame and insert each row
    for _, row in df_countries.iterrows():
        cursor.execute(insert_stmt, tuple(row))

    # Commit the transaction
    connection.commit()

    print("transformed_countries_dataset.csv inserted into MySQL successfully.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()

df_genres = pd.read_csv('../1_Data/transformed_genres_dataset.csv')

try:
    connection = mysql.connector.connect(**configuration)
    cursor = connection.cursor()

    cursor.execute("USE netflix_titles;")
    
    insert_stmt = """
        INSERT INTO genres (title, genre)
        VALUES (%s, %s)
    """

    # Iterate through the DataFrame and insert each row
    for _, row in df_genres.iterrows():
        cursor.execute(insert_stmt, tuple(row))

    # Commit the transaction
    connection.commit()

    print("transformed_genres_dataset.csv inserted into MySQL successfully.")

    cursor.close()
    connection.close()

except mysql.connector.Error as err:
    if err.errno == errorcode.ER_ACCESS_DENIED_ERROR:
        print("Something is wrong with your user name or password")
    elif err.errno == errorcode.ER_BAD_DB_ERROR:
        print("Database does not exist")
    else:
        print(err)
else:
    connection.close()
